requirements.txt
datasets==2.9.0
diffusers==0.12.1
dlib==19.24.1
einops==0.6.0
evaluate==0.4.0
h5py==3.8.0
ipython==8.12.0
joblib==1.2.0
lmdb==1.4.1
lpips==0.1.4
matplotlib==3.6.3
natsort==8.2.0
numpy==1.23.5
opencv_python==4.7.0.68
Pillow==9.5.0
PyYAML==6.0
requests==2.28.1
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.10.1
sentence_transformers==2.2.2
setuptools==65.6.3
tqdm==4.64.1
transformers==4.28.0
undecorated==0.3.0
wandb==0.14.2
langchain==0.0.153
torchmetrics==1.2.1
bert-score==0.3.13
timm==0.9.12
openai==1.3.7
accelerate==0.23.0
